Adding base class for uevent and umockdev device manager
