+++ src/Library/UMockdevDeviceManager.hpp
@@ -22,105 +22,27 @@
 #endif
 
 #if defined(HAVE_UMOCKDEV)
-
 #include "Common/Thread.hpp"
-#include "SysFSDevice.hpp"
+#include "DeviceManagerBase.hpp"
 #include "UMockdevDeviceDefinition.hpp"
 
-#include "usbguard/Typedefs.hpp"
-#include "usbguard/DeviceManager.hpp"
-#include "usbguard/Device.hpp"
-#include "usbguard/Rule.hpp"
-#include "usbguard/USB.hpp"
-
 #include <condition_variable>
-#include <istream>
-
-#include <sys/stat.h>
-#include <dirent.h>
-
 #include <umockdev.h>
 
 namespace usbguard
 {
-  class UMockdevDeviceManager;
+  class UEvent;
 
-  class UMockdevDevice : public Device, public USBDescriptorParserHooks
+  class UMockdevDeviceManager : public DeviceManagerBase
   {
-  public:
-    UMockdevDevice(UMockdevDeviceManager& device_manager, SysFSDevice& sysfs_device);
-
-    SysFSDevice& sysfsDevice();
-    const std::string& getSysPath() const;
-    bool isController() const override;
-    std::string getSystemName() const override;
-
-  private:
-    void parseUSBDescriptor(USBDescriptorParser* parser, const USBDescriptor* descriptor_raw,
-      USBDescriptor* descriptor_out) override;
-    void loadUSBDescriptor(USBDescriptorParser* parser, const USBDescriptor* descriptor) override;
-    bool isLinuxRootHubDeviceDescriptor(const USBDescriptor* descriptor);
-    void updateHashLinuxRootHubDeviceDescriptor(const USBDescriptor* descriptor);
-
-    SysFSDevice _sysfs_device;
-  };
-
-  /*
-   * TODO: Create a base class template that provides
-   *       a shared implementation for Linux based device manager.
-   */
-  class UMockdevDeviceManager : public DeviceManager
-  {
-    using DeviceManager::insertDevice;
-
   public:
     UMockdevDeviceManager(DeviceManagerHooks& hooks);
     ~UMockdevDeviceManager();
-
-    void setEnumerationOnlyMode(bool state) override;
 
     void start() override;
     void stop() override;
     void scan() override;
     void scan(const std::string& devpath) override;
-
-    std::shared_ptr<Device> applyDevicePolicy(uint32_t id, Rule::Target target) override;
-    void insertDevice(std::shared_ptr<UMockdevDevice> device);
-    std::shared_ptr<Device> removeDevice(const std::string& syspath);
-
-    uint32_t getIDFromSysfsPath(const std::string& sysfs_path) const;
-
-  protected:
-    void umockdevInit();
-    std::vector<std::string> umockdevLoadFromFile(const std::string& definitions_path);
-    std::vector<std::string> umockdevRemoveByFile(const std::string& definitions_path);
-    void umockdevRemoveBySysfsPath(const std::string& sysfs_path);
-    void umockdevProcessInotify();
-    std::vector<std::string> umockdevGetChildrenBySysfsPath(const std::string& sysfs_path);
-    void umockdevAuthorizeBySysfsPath(const std::string& sysfs_path);
-    void umockdevDeauthorizeBySysfsPath(const std::string& sysfs_path);
-    void umockdevAdd(const std::shared_ptr<UMockdevDeviceDefinition>& definition);
-    void umockdevRemove(const std::shared_ptr<UMockdevDeviceDefinition>& definition);
-    void umockdevRemove(const std::string& sysfs_path);
-
-    int ueventOpen();
-    void sysfsApplyTarget(SysFSDevice& sysfs_device, Rule::Target target);
-
-    void thread();
-    void ueventProcessRead();
-    void ueventProcessUEvent(const UEvent& uevent);
-    static bool ueventEnumerateComparePath(const std::pair<std::string, std::string>& a,
-      const std::pair<std::string, std::string>& b);
-    int ueventEnumerateDevices();
-
-    static std::string ueventEnumerateFilterDevice(const std::string& filepath, const struct dirent* direntry);
-    int ueventEnumerateTriggerAndWaitForDevice(const std::string& devpath, const std::string& buspath);
-
-    void processDevicePresence(SysFSDevice& sysfs_device);
-
-    void processDeviceInsertion(SysFSDevice& sysfs_device, bool signal_present);
-    void processDevicePresence(uint32_t id);
-    void processDeviceRemoval(const std::string& sysfs_devpath);
 
   private:
     struct GObjectDeleter {
@@ -132,31 +54,35 @@ namespace usbguard
       }
     };
 
+    static bool ueventEnumerateComparePath(const std::pair<std::string, std::string>& a,
+      const std::pair<std::string, std::string>& b);
+
+    void umockdevInit();
+    void umockdevAdd(const std::shared_ptr<UMockdevDeviceDefinition>& definition);
+    void umockdevRemove(const std::shared_ptr<UMockdevDeviceDefinition>& definition);
+    void umockdevRemove(const std::string& sysfs_path);
+    void umockdevProcessInotify();
+    std::vector<std::string> umockdevLoadFromFile(const std::string& definitions_path);
+    std::vector<std::string> umockdevRemoveByFile(const std::string& definitions_path);
+    std::vector<std::string> umockdevGetChildrenBySysfsPath(const std::string& sysfs_path);
+
+    void sysfsAuthorizeDevice(SysFSDevice& sysfs_device) override;
+    void sysfsDeauthorizeDevice(SysFSDevice& sysfs_device) override;
+
+    void thread();
+    void ueventProcessRead();
+    void ueventProcessUEvent(const UEvent& uevent);
+    int ueventEnumerateDevices();
+    int ueventEnumerateTriggerAndWaitForDevice(const std::string& devpath, const std::string& buspath);
+
     Thread<UMockdevDeviceManager> _thread;
     std::unique_ptr<UMockdevTestbed, GObjectDeleter> _testbed{nullptr};
     std::string _umockdev_deviceroot;
     int _inotify_fd{-1};
-    int _uevent_fd{-1};
-    int _wakeup_fd{-1};
-
-    /*
-     * The following maps sysfs devices paths to their IDs.
-     * The key must not contain the sysfs (/sys) directory
-     * root prefix in the path and must be normalized to not
-     * contain ./ and ../ path components.
-     */
-    std::map<std::string, uint32_t> _sysfs_path_to_id_map;
-
-    bool isPresentSysfsPath(const std::string& sysfs_path) const;
-    bool knownSysfsPath(const std::string& sysfs_path, uint32_t* id = nullptr) const;
-    void learnSysfsPath(const std::string& sysfs_path, uint32_t id = 0);
-    void forgetSysfsPath(const std::string& sysfs_path);
 
     std::map<std::string, std::shared_ptr<UMockdevDeviceDefinition>> _sysfs_path_map;
     std::multimap<std::string, std::weak_ptr<UMockdevDeviceDefinition>> _umockdev_file_map;
 
-    bool _enumeration_only_mode{false};
-    std::atomic<bool> _enumeration{false};
     std::condition_variable _enumeration_complete;
     std::mutex _enumeration_mutex;
   };
